refactor: common이 아닌 컴포넌트 위치 변경
