+++ src/components/ProductCard/ProductCard.tsx
@@ -1,7 +1,7 @@
 import { memo } from "react";
 import { ReactComponent as ShoppingCartIcon } from "../../assets/icon/shopping-cart.svg";
 import Counter from "../common/Counter/Counter";
-import ProductImg from "../common/ProductImg/ProductImg";
+import ProductImg from "../ProductImg/ProductImg";
 import ProductInfo from "./ProductInfo/ProductInfo";
 import useCart from "../../hooks/useCart";
 import Styled from "./ProductCardStyled";
